separate inference partnership

+++ inference.py
@@ -90,9 +90,9 @@ def main():
                 mask_positions.append(i)
 
         output = model_fn.forward(
-            input_ids = torch.FloatTensor(encoded_sequence) , 
+            input_ids = torch.FloatTensor(encoded_sequence), 
             attention_mask = padded_sequences["attention_mask"],
-            mask_pos = mask_positions,
+            mask_pos = torch.FloatTensor(mask_positions),
             labels = ['yes', 'no'])
         
         print(output)
